Clarify outside employment during leave
Our current policy is way to strict about outside work while on leave.

We’ve generally taken a relaxed approach to this as long as it is not
competitive or conflicting, and if you help someone out for a few days or week, one or
twice, to make some extra cash for rent, that's no big deal to
us, assuming your plan is to return to thoughtbot at some point.

Generally, this should match our existing policy regarding outside work,
and so I've linked that policy instead and tried to clarify there.