Update overtime section
The current policy was copied from an old employee handbook provided by
Paychex, it does not reflect the fact that we really only have exempt
employees, and that overtime is extremely rare.

We've had a practice of comp-time, which is not at all required for
exempt employees, but we do it for sustainable pace reasons. This change
actually includes that as an official thing.